+++ packages/flow-lineage/CHANGELOG.md
@@ -1,5 +1,14 @@
 <h4 class="margin-btm-8">Release Notes</h4>
 
+## 3.4.2
+
+### Patch Changes
+
+- build cleanup
+- Updated dependencies
+  - @nonfx/flow-core-config@1.3.3
+  - @nonfx/flow-core@6.0.2
+
 ## 3.4.1
 
 ### Patch Changes